conditional compilation for windows-only stuff
