+++ backend/src/result.rs
@@ -1,10 +1,10 @@
-use std::{
-	path::PathBuf,
-	result,
-};
+use std::{path::PathBuf, result};
 
 #[derive(Debug, thiserror::Error)]
 pub enum Error {
+	#[error("Not implemented")]
+	NotImplemented,
+
 	#[error(transparent)]
 	Io(#[from] std::io::Error),
