+++ frontend/api/bindings.ts
@@ -62,16 +62,16 @@ export function openModsFolder() {
     return invoke()<null>("open_mods_folder")
 }
 
-export type GameEngineVersion = { major: number; minor: number; patch: number; suffix: string | null; display: string }
-export type GameEngineBrand = "Unity" | "Unreal" | "Godot"
 export type DiscoverGame = { id: string; engine: GameEngineBrand; nsfw: boolean }
-export type ModKind = "Installable" | "Runnable"
-export type OwnedGame = { id: string; name: string; installed: boolean; osList: OperatingSystem[]; engine: GameEngineBrand; releaseDate: number; thumbnailUrl: string }
-export type Game = { id: string; name: string; discriminator: string | null; fullPath: string; architecture: Architecture | null; scriptingBackend: UnityScriptingBackend | null; operatingSystem: OperatingSystem | null; steamLaunch: SteamLaunchOption | null; installedMods: string[]; engine: GameEngine | null; thumbnailUrl: string | null }
-export type SteamLaunchOption = { launchId: string; appId: number; description: string | null; executable: string | null; arguments: string | null; appType: string | null; osList: string | null; betaKey: string | null; osArch: string | null }
-export type Mod = { id: string; name: string; scriptingBackend: UnityScriptingBackend | null; engine: GameEngineBrand | null; kind: ModKind; path: string }
-export type UnityScriptingBackend = "Il2Cpp" | "Mono"
-export type Architecture = "X64" | "X86"
-export type ModLoaderData = { id: string; path: string; mods: Mod[] }
-export type OperatingSystem = "Linux" | "Windows"
 export type GameEngine = { brand: GameEngineBrand; version: GameEngineVersion | null }
+export type Mod = { id: string; name: string; scriptingBackend: UnityScriptingBackend | null; engine: GameEngineBrand | null; kind: ModKind; path: string }
+export type ModLoaderData = { id: string; path: string; mods: Mod[] }
+export type SteamLaunchOption = { launchId: string; appId: number; description: string | null; executable: string | null; arguments: string | null; appType: string | null; osList: string | null; betaKey: string | null; osArch: string | null }
+export type GameEngineVersion = { major: number; minor: number; patch: number; suffix: string | null; display: string }
+export type UnityScriptingBackend = "Il2Cpp" | "Mono"
+export type OwnedGame = { id: string; name: string; installed: boolean; osList: OperatingSystem[]; engine: GameEngineBrand; releaseDate: number; thumbnailUrl: string }
+export type ModKind = "Installable" | "Runnable"
+export type Architecture = "X64" | "X86"
+export type GameEngineBrand = "Unity" | "Unreal" | "Godot"
+export type Game = { id: string; name: string; discriminator: string | null; fullPath: string; architecture: Architecture | null; scriptingBackend: UnityScriptingBackend | null; operatingSystem: OperatingSystem | null; steamLaunch: SteamLaunchOption | null; installedMods: string[]; engine: GameEngine | null; thumbnailUrl: string | null }
+export type OperatingSystem = "Linux" | "Windows"
